src/poise/requirements.txt
textual==0.86.2
httpx==0.27.0
pydantic>=2.7.0
python-dotenv==1.0.0
py-clob-client==0.19.0
tzdata>=2024.1
